refactoring, restructuring Player, Players
