+++ core/src/space/hypeo/networking/IPlayerConnector.java
@@ -1,7 +1,6 @@
 package space.hypeo.networking;
 
-import java.util.HashMap;
-import java.util.List;
+import space.hypeo.networking.packages.Players;
 
 
 /**
